// src/api/generated/models/getCompaniesCompanyIdRequestLinesInboundParams.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */
import type { GetCompaniesCompanyIdRequestLinesInboundStatus } from './getCompaniesCompanyIdRequestLinesInboundStatus';
import type { GetCompaniesCompanyIdRequestLinesInboundWithoutQuotes } from './getCompaniesCompanyIdRequestLinesInboundWithoutQuotes';
import type { GetCompaniesCompanyIdRequestLinesInboundSortBy } from './getCompaniesCompanyIdRequestLinesInboundSortBy';
import type { GetCompaniesCompanyIdRequestLinesInboundSortOrder } from './getCompaniesCompanyIdRequestLinesInboundSortOrder';

export type GetCompaniesCompanyIdRequestLinesInboundParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
/**
 * @minimum 0
 * @maximum 9007199254740991
 */
offset?: number;
status?: GetCompaniesCompanyIdRequestLinesInboundStatus;
/**
 * @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
 */
requestId?: string;
/**
 * @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
 */
productId?: string;
/**
 * @minLength 1
 */
q?: string;
withoutQuotes?: GetCompaniesCompanyIdRequestLinesInboundWithoutQuotes;
sortBy?: GetCompaniesCompanyIdRequestLinesInboundSortBy;
sortOrder?: GetCompaniesCompanyIdRequestLinesInboundSortOrder;
};